Move InvalidConfigurationException to infrastructure:exceptions
