+++ data/beaconrestapi/build.gradle
@@ -2,9 +2,10 @@ dependencies {
     implementation project(':data')
     implementation project(':data:provider')
     implementation project(':data:serializer')
-implementation project(':ethereum:spec')
+    implementation project(':ethereum:spec')
     implementation project(':ethereum:statetransition')
     implementation project(':infrastructure:async')
+    implementation project(':infrastructure:exceptions')
     implementation project(':infrastructure:http')
     implementation project(':infrastructure:logging')
     implementation project(':infrastructure:version')
