+++ infrastructure/exceptions/src/main/java/tech/pegasys/teku/infrastructure/exceptions/InvalidConfigurationException.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.config;
+package tech.pegasys.teku.infrastructure.exceptions;
 
 public class InvalidConfigurationException extends RuntimeException {
   public InvalidConfigurationException(final String message) {
