+++ services/beaconchain/build.gradle
@@ -10,6 +10,7 @@ dependencies {
   implementation project(':ethereum:statetransition')
   implementation project(':ethereum:weaksubjectivity')
   implementation project(':infrastructure:async')
+  implementation project(':infrastructure:exceptions')
   implementation project(':infrastructure:http')
   implementation project(':infrastructure:logging')
   implementation project(':infrastructure:version')
