+++ services/serviceutils/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
   implementation project(':infrastructure:metrics')
   implementation project(':infrastructure:events')
+  implementation project(':infrastructure:exceptions')
   implementation project(':infrastructure:async')
   implementation project(':infrastructure:logging')
   implementation project(':util')
