+++ teku/src/main/java/tech/pegasys/teku/cli/subcommand/PeerCommand.java
@@ -35,7 +35,7 @@ import picocli.CommandLine.ParentCommand;
 import picocli.CommandLine.Spec;
 import tech.pegasys.teku.cli.BeaconNodeCommand;
 import tech.pegasys.teku.cli.converter.PicoCliVersionProvider;
-import tech.pegasys.teku.util.config.InvalidConfigurationException;
+import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
 
 @Command(
     name = "peer",
