+++ util/src/main/java/tech/pegasys/teku/util/config/ConstantsReader.java
@@ -29,12 +29,13 @@ import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import org.apache.tuweni.bytes.Bytes;
+import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
 import tech.pegasys.teku.infrastructure.io.resource.ResourceLoader;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.ssz.type.Bytes4;
 
 class ConstantsReader {
-  private static ImmutableList<String> PRESETS = ImmutableList.of("mainnet", "minimal");
+  private static final ImmutableList<String> PRESETS = ImmutableList.of("mainnet", "minimal");
   private static final String PRESET_PATH = "presets/";
   private static final String CONFIG_PATH = "configs/";
   private static final String PRESET_FIELD = "PRESET_BASE";
