+++ validator/api/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
   implementation project(':bls')
   implementation project(':infrastructure:async')
+  implementation project(':infrastructure:exceptions')
   implementation project(':ethereum:core')
   implementation project(':ethereum:spec')
   implementation project(':data:serializer')
