+++ validator/api/src/main/java/tech/pegasys/teku/validator/api/InteropConfig.java
@@ -15,9 +15,9 @@ package tech.pegasys.teku.validator.api;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
+import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.config.SpecConfig;
-import tech.pegasys.teku.util.config.InvalidConfigurationException;
 
 public class InteropConfig {
