+++ validator/api/src/main/java/tech/pegasys/teku/validator/api/KeyStoreFilesLocator.java
@@ -30,7 +30,7 @@ import java.util.stream.Stream;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import tech.pegasys.teku.util.config.InvalidConfigurationException;
+import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
 
 class KeyStoreFilesLocator {
   private static final Logger LOG = LogManager.getLogger();
