+++ validator/client/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
   api 'org.bouncycastle:bcprov-jdk15on'
   implementation project(':bls')
+  implementation project(':infrastructure:exceptions')
   implementation project(':infrastructure:io')
   implementation project(':infrastructure:metrics')
   implementation project(':infrastructure:http')
