+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/loader/LocalValidatorSource.java
@@ -39,8 +39,8 @@ import tech.pegasys.teku.bls.BLSSecretKey;
 import tech.pegasys.teku.core.signatures.LocalSigner;
 import tech.pegasys.teku.core.signatures.Signer;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
+import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
 import tech.pegasys.teku.spec.Spec;
-import tech.pegasys.teku.util.config.InvalidConfigurationException;
 import tech.pegasys.teku.validator.api.ValidatorConfig;
 
 public class LocalValidatorSource implements ValidatorSource {
