+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/loader/PublicKeyLoader.java
@@ -24,7 +24,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.tuweni.bytes.Bytes;
 import tech.pegasys.teku.bls.BLSPublicKey;
-import tech.pegasys.teku.util.config.InvalidConfigurationException;
+import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
 
 public class PublicKeyLoader {
   final ObjectMapper objectMapper;
