+++ validator/client/src/test/java/tech/pegasys/teku/validator/client/loader/PublicKeyLoaderTest.java
@@ -26,8 +26,8 @@ import java.util.List;
 import org.apache.tuweni.bytes.Bytes48;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.bls.BLSPublicKey;
+import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
-import tech.pegasys.teku.util.config.InvalidConfigurationException;
 
 public class PublicKeyLoaderTest {
   private final DataStructureUtil dataStructureUtil = new DataStructureUtil();
